+++ package.json
@@ -7,7 +7,9 @@
     "packages/common"
   ],
   "scripts": {
-    "build": "yarn workspace @packages/astro build && yarn workspace @packages/common build"
+    "build": "yarn build:common & yarn build:astro",
+    "build:astro": "yarn workspace @packages/astro build",
+    "build:common": "yarn workspace @packages/common build"
   },
   "dependencies": {
     "@packages/astro": "*",
