+++ packages/astro/package.json
@@ -12,6 +12,7 @@
     "astro": "astro"
   },
   "dependencies": {
+    "@packages/common": "*",
     "@astrojs/node": "^8.2.5",
     "astro": "^4.5.12"
   }
